Clean up, structural update to meet 'production standarts'
