--- app/bak/main copy.py
@@ -1,108 +1,0 @@
-from flask import Flask, render_template, redirect, url_for, request, session
-from flask_sqlalchemy import SQLAlchemy
-from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user, current_user
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, IntegerField, SubmitField
-from wtforms.validators import DataRequired, Email
-from werkzeug.security import generate_password_hash, check_password_hash
-from datetime import datetime
-
-app = Flask(__name__)
-app.config['SECRET_KEY'] = 'secret'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///data.db'
-db = SQLAlchemy(app)
-
-login_manager = LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = 'login'
-
-### MODELS ###
-class User(UserMixin, db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password_hash = db.Column(db.String(128), nullable=False)
-    cars = db.relationship('Car', backref='owner', lazy=True)
-
-class Car(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    make = db.Column(db.String(64), nullable=False)
-    model = db.Column(db.String(64), nullable=False)
-    year = db.Column(db.Integer)
-    color = db.Column(db.String(32))
-    vin = db.Column(db.String(128))
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-
-### FORMS ###
-class LoginForm(FlaskForm):
-    email = StringField('Email', validators=[DataRequired(), Email()])
-    password = PasswordField('Password', validators=[DataRequired()])
-    submit = SubmitField('Login')
-
-class CarForm(FlaskForm):
-    make = StringField('Make', validators=[DataRequired()])
-    model = StringField('Model', validators=[DataRequired()])
-    year = IntegerField('Year')
-    color = StringField('Color')
-    vin = StringField('VIN')
-    submit = SubmitField('Save')
-
-### LOGIN LOADER ###
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
-
-### ROUTES ###
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and check_password_hash(user.password_hash, form.password.data):
-            login_user(user)
-            return redirect(url_for('car_list'))
-    return render_template('login.html', form=form)
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('login'))
-
-@app.route('/cars')
-@login_required
-def car_list():
-    cars = Car.query.filter_by(user_id=current_user.id).all()
-    return render_template('car_list.html', cars=cars)
-
-@app.route('/cars/add', methods=['GET', 'POST'])
-@login_required
-def add_car():
-    form = CarForm()
-    if form.validate_on_submit():
-        car = Car(**form.data, user_id=current_user.id)
-        db.session.add(car)
-        db.session.commit()
-        return redirect(url_for('car_list'))
-    return render_template('add_car.html', form=form)
-
-@app.route('/cars/delete/<int:car_id>', methods=['POST'])
-@login_required
-def delete_car(car_id):
-    car = Car.query.get_or_404(car_id)
-    if car.user_id != current_user.id:
-        return "Unauthorized", 403
-    db.session.delete(car)
-    db.session.commit()
-    return redirect(url_for('car_list'))
-
-### INIT ###
-if __name__ == '__main__':
-    with app.app_context():
-        db.create_all()
-        # Create a dummy user for testing if none exists
-        if not User.query.filter_by(email='test@example.com').first():
-            user = User(email='test@example.com', password_hash=generate_password_hash('123'))
-            db.session.add(user)
-            db.session.commit()
-    app.run(debug=True)
