--- app/forms/test_form.py
@@ -1,6 +1,0 @@
-from flask_wtf import FlaskForm
-from wtforms import SubmitField
-
-
-class TestForm(FlaskForm):
-    submit = SubmitField('Test select_where')
