+++ app/models/__init__.py
@@ -1,0 +1,26 @@
+
+
+
+
+from app.models.db.user import User
+from app.models.db.faculty import Faculty
+from app.models.db.student_group import StudentGroup
+from app.models.db.study_program import StudyProgram
+from app.models.db.module import Module
+from app.models.db.enrollment import Enrollment
+from app.models.db.assessment import Assessment
+
+from app.models.db.car import Car
+
+
+
+# __all__ = [
+#     "User",
+#     "Faculty",
+#     "StudentGroup",
+#     "StudyProgram",
+#     "Module",
+#     "Enrollment",
+#     "Assessment",
+#     "Car"
+# ]
