+++ app/models/db/faculty.py
@@ -1,7 +1,10 @@
-from db import db
+from app import db
 from sqlalchemy.orm import relationship
 
+
 class Faculty(db.Model):
+    __tablename__ = 'faculty'
+
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False)
