+++ app/models/db/user.py
@@ -56,15 +56,3 @@ class User(UserMixin, db.Model):
         remaining = self.blocked_until - datetime.utcnow()
         return max(0, int(remaining.total_seconds() / 60))
  
-
-
-class StudentGroup(db.Model):
-    __tablename__ = 'student_group'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(20), nullable=False)
-
-    program_id = db.Column(db.Integer, db.ForeignKey('study_program.id', ondelete='SET NULL'), nullable=True)
-    program = relationship("StudyProgram", back_populates="groups", passive_deletes=True)
-
-    students = relationship("User", back_populates="group")
