--- app/models/old/user_old.py
@@ -1,12 +1,0 @@
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.orm import relationship
-from flask_login import UserMixin
-from app import db
-
-
-class User(UserMixin, db.Model):
-    id = Column(Integer, primary_key=True)
-    username = Column(String(64), unique=True, nullable=False)
-    email = Column(String(120), unique=True, nullable=False)
-    password_hash = Column(String(128), nullable=False)
-    cars = relationship('Car', backref='owner', lazy=True)
